refactor(lang): move materializeLowered from unplugin to lang
The `.uwk.ts` multi-file lowering routine (recursive lower → temp
sibling → cleanup path list) was private to
packages/unplugin/src/index.ts, so callers outside the Vite plugin
(offline, tests, headless authoring) had no access to it. Pull it up
to the package the downstream dependency direction already points at
(unplugin → lang), so a follow-up commit can add a `loadUwkProcessor`
convenience wrapper that lets `renderOffline` render a multi-file
`.uwk.ts` processor without a bundler.

Behavior change: zero. The unplugin runtime is byte-identical — same
functions, imported from @unworklet/lang instead of defined locally.
- The shared lowered-source cache stays in one place (now inside
  materialize-lowered.ts) so both callers benefit.
- `isUwkSource`, `deriveExportName`, `lowerUwkSource`, and
  `materializeLowered` are the four exports; the cyclic-import error
  string is updated from `@unworklet/unplugin` to `@unworklet/lang` to
  match the new owner.

Setup for the next commit (feat + test) that adds `loadUwkProcessor`
and an offline test rendering a multi-file `.uwk.ts`.
